+++ R/meta_estimator.R
@@ -5,6 +5,8 @@
 #' @param which_estimand name of study level estimand for meta analysis
 #' @param benchmark named list of length 2 giving benchmark sampling-estimator pair (only accepts one value across studies for now)
 #' @param stan_handler function that takes stan_data as input and produces compilable stan model object
+#' @param hidden_prior list of two hyperpriors, on means and standard errors of each included sampling-estimator pairs. Names of list objects should be "mean" and "se". If one number provided for a hyperprior it gets expanded to all sampling-estimator pairs
+#' @param rel_bias_prior list of two hyperpriors, on means and standard errors of relative bias. Names of list objects should be "mean" and "se".
 #' @param control_params list of additional parameters to pass to Stan fit function. These can include number of iterations, chains, thinning, seed and number of cores to use
 #'
 #' @return
@@ -18,7 +20,9 @@ get_meta_estimates <- function(
   sampling_variable = "meta",
   which_estimand = "hidden_size",
   benchmark = list(sample = "pps", estimator = "ht"),
-  stan_handler = get_meta_stan2,
+  stan_handler = get_meta_stan3,
+  hidden_prior = NULL,
+  rel_bias_prior = list(mean = 1, se = 10),
   control_params = list(
     iter = 8000, chains = 8, thin = 10,
     seed = 872312,
@@ -27,7 +31,7 @@ get_meta_estimates <- function(
 
   .stan_data <-
     data %>%
-    dplyr::filter(across(all_of(sampling_variable), ~ . == 1),
+    dplyr::filter(dplyr::if_any(sampling_variable, ~ . == 1),
                   inquiry %in% which_estimand)
 
   .samp_ests <- unique(.stan_data[,c("sample", "estimator")])
@@ -44,6 +48,28 @@ get_meta_estimates <- function(
 
   .N <- length(.studies)
   .K <- nrow(.samp_ests)
+
+  if (!is.null(hidden_prior) & is.list(hidden_prior)) {
+    .alpha_prior <- do.call(cbind, hidden_prior)
+
+    if (nrow(.alpha_prior) == 1) {
+      .alpha_prior <- .alpha_prior[rep(1, times = .N),]
+    } else if (nrow(.alpha_prior) == .N) {
+      .alpha_prior <- .alpha_prior[.studies,]
+    } else {
+      stop("There is mismatch in length of hyperpriors on mean and std. error of target parameters. Length of each element should be equal either to length of number of studies included or to 1")
+    }
+  } else if (is.null(hidden_prior)) {
+    .alpha_prior <-
+      .stan_data %>%
+      group_by(study) %>%
+      summarize(mean = sum(estimate / (se * sum(1/se))), se = 2 * mean(se / (se * sum(1/se)))) %>%
+      { .[order(match(.$study,.studies)),] } %>%
+      select(-study) %>%
+      as.matrix()
+  } else {
+    stop("Hyperpriors on target parameters are provided in wrong format (should be NULL or list of two numeric objects)")
+  }
 
   # get ids of unique samp-est pairs in observed data
   .samp_est_ids <-
@@ -68,9 +94,12 @@ get_meta_estimates <- function(
       .,
       apply(X = expand.grid(1:.K, c("n", "observed", "est", "est_se")),
             MARGIN = 1,
-            FUN = function(x) paste0(x[c(2,1)], collapse = ""))
+            FUN = function(x) paste0(x[2], as.integer(x[1])))
     ) %>%
-    c(N = .N, K = .K, .)
+    c(list(N = .N, K = .K,
+           alpha_mean = unname(.alpha_prior[,"mean"]), alpha_se = unname(.alpha_prior[,"se"]),
+           rel_bias_mean = rel_bias_prior$mean, rel_bias_se = rel_bias_prior$se),
+      .)
 
   .stan_model <-
     rstan::stan_model(model_code = stan_handler(.stan_data))
@@ -119,8 +148,12 @@ get_meta_estimates <- function(
                se =   c(unname(.biases[[2]][1,]), unname(.study_ests[,2])),
                inquiry = c(paste0("rel_bias_", .samp_est_names),
                            paste0(.studies, "_", which_estimand)),
+               # note assumes first item is benchmark
+               prior = c(1, rep(rel_bias_prior[[1]], .K-1), .alpha_prior[,1]) %>% unlist,
+               prior_sd = c(0, rep(rel_bias_prior[[2]], .K-1), .alpha_prior[,2]) %>% unlist,
                stringsAsFactors = FALSE
     )
   )
 
 }
+